POPULATION DATA ANALYSIS PROJECT USING EXCEL
The purpose of this analysis was to identify the factors affecting or influencing population growth. Based on the analysis, it can be concluded that Migrant Rate and Density are related to each other. If Migrant Rate increases, Density also increases and vice versa. Similarly Fertility Rate and Yearly % Change are also inter-related.